requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
python-dotenv>=1.0
psycopg[binary]>=3.2
psycopg-pool>=3.2
pydantic-settings>=2.0
httpx>=0.27
Pillow>=10.0
python-multipart>=0.0.9
